requirements.txt
# Database
pyodbc>=5.1.0
sqlalchemy>=2.0.0

# Config
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# API Server
fastapi>=0.110.0
uvicorn>=0.27.0

# LLM
openai>=1.10.0
httpx>=0.27.0

# Agent Framework (XiYan auto-description pipeline)
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0

# Vector DB (embeddings via OpenAI API in askdataai/indexing/embedder.py)
chromadb>=0.5.0
numpy

# Security
sqlparse>=0.5.0

# Stage 0 Guardrail: PIGuard (leolee99/PIGuard — Prompt Injection detection, ACL 2025)
# Model: DeBERTa-v3-base fine-tuned for prompt injection classification
transformers>=4.40.0
torch>=2.0.0
sentencepiece>=0.1.99

# Stage 0.5 Conversation Context: mem0 (open-source, ChromaDB local)
mem0ai>=0.1.0
